finalizing updated detect-strand
